# backend/UteLearningHub.Api/Controllers/CommentController.cs
using MediatR;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.RateLimiting;
using UteLearningHub.Application.Common.Dtos;
using UteLearningHub.Application.Features.Comment.Commands.CreateComment;
using UteLearningHub.Application.Features.Comment.Commands.DeleteComment;
using UteLearningHub.Application.Features.Comment.Commands.UpdateComment;
using UteLearningHub.Application.Features.Comment.Queries.GetComments;
using UteLearningHub.Application.Features.Comment.Queries.GetDocumentComments;

namespace UteLearningHub.Api.Controllers;

[Route("api/[controller]")]
[ApiController]
public class CommentController : ControllerBase
{
    private readonly IMediator _mediator;

    public CommentController(IMediator mediator)
    {
        _mediator = mediator;
    }

    [HttpGet]
    public async Task<ActionResult<PagedResponse<CommentDetailDto>>> GetComments([FromQuery] GetCommentsQuery query)
    {
        var result = await _mediator.Send(query);
        return Ok(result);
    }

    [HttpGet("document")]
    public async Task<ActionResult<PagedResponse<CommentDetailDto>>> GetDocumentComments([FromQuery] GetDocumentCommentsQuery query)
    {
        var result = await _mediator.Send(query);
        return Ok(result);
    }

    [HttpPost]
    [Authorize]
    [EnableRateLimiting("comment")]
    public async Task<ActionResult<CommentDetailDto>> CreateComment([FromBody] CreateCommentCommand command)
    {
        var result = await _mediator.Send(command);
        return Ok(result);
    }

    [HttpPut("{id}")]
    [Authorize]
    public async Task<ActionResult<CommentDetailDto>> UpdateComment(Guid id, [FromBody] UpdateCommentCommandRequest request)
    {
        var command = new UpdateCommentCommand
        {
            Id = id,
            Content = request.Content
        };
        var result = await _mediator.Send(command);
        return Ok(result);
    }

    [HttpDelete("{id}")]
    [Authorize]
    public async Task<IActionResult> DeleteComment(Guid id)
    {
        var command = new DeleteCommentCommand { Id = id };
        await _mediator.Send(command);
        return NoContent();
    }
}