# frontend/openapi/database_api_service.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "UteLearningHub.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:7080/"
    }
  ],
  "paths": {
...
    "/api/Comment": {
      "get": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "name": "DocumentFileId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ParentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfCommentDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfCommentDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfCommentDetailDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Comment"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Comment/document": {
      "get": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "name": "DocumentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "ParentId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfCommentDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfCommentDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfCommentDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Comment/{id}": {
      "put": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentCommandRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentCommandRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCommentCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDetailDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Comment"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Document": {
      "get": {
        "tags": [
          "Document"
        ],
        "parameters": [
          {
            "name": "SubjectId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "TypeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "TagIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "MajorId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AuthorId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "CreatedById",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "SearchTerm",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Visibility",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDescending",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "IsDeleted",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfDocumentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfDocumentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponseOfDocumentDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Document"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Document/{id}": {
      "get": {
        "tags": [
          "Document"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Document"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentCommandRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentCommandRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Document"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "documentCount": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean",
            "nullable": true
          },
          "deletedById": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "createdById": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "updatedById": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "AuthorDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "AuthorInput": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "CommentDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "documentFileId": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "content": {
            "type": "string"
          },
          "authorName": {
            "type": "string"
          },
          "authorAvatarUrl": {
            "type": "string",
            "nullable": true
          },
          "createdById": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus"
          },
          "replyCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "ContentStatus": {
        "type": "integer"
      },
...
      "CreateCommentCommand": {
        "type": "object",
        "properties": {
          "documentFileId": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "content": {
            "type": "string"
          }
        }
      },
...
      "CreateDocumentCommand": {
        "type": "object",
        "properties": {
          "documentName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "subjectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "typeId": {
            "type": "string",
            "format": "uuid"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "tagNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "authorIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorInput"
            },
            "nullable": true
          },
          "visibility": {
            "$ref": "#/components/schemas/VisibilityStatus"
          },
          "coverFileId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
...
      "DocumentDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "documentName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/VisibilityStatus"
          },
          "subject": {
            "$ref": "#/components/schemas/SubjectDto"
          },
          "type": {
            "$ref": "#/components/schemas/TypeDto"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDto"
            }
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorDto"
            }
          },
          "coverFileId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentFileDto"
            }
          },
          "commentCount": {
            "type": "integer",
            "format": "int32"
          },
          "usefulCount": {
            "type": "integer",
            "format": "int32"
          },
          "notUsefulCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalViewCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdById": {
            "type": "string",
            "format": "uuid"
          },
          "createdByName": {
            "type": "string",
            "nullable": true
          },
          "createdByAvatarUrl": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "DocumentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "documentName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/VisibilityStatus"
          },
          "subject": {
            "$ref": "#/components/schemas/SubjectDto"
          },
          "type": {
            "$ref": "#/components/schemas/TypeDto"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagDto"
            }
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorDetailDto"
            }
          },
          "thumbnailFileId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "fileCount": {
            "type": "integer",
            "format": "int32"
          },
          "commentCount": {
            "type": "integer",
            "format": "int32"
          },
          "usefulCount": {
            "type": "integer",
            "format": "int32"
          },
          "notUsefulCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalViewCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdById": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DocumentFileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileId": {
            "type": "string",
            "format": "uuid"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "mimeType": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "order": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "coverFileId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus"
          },
          "reviewedById": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "reviewNote": {
            "type": "string",
            "nullable": true
          },
          "commentCount": {
            "type": "integer",
            "format": "int32"
          },
          "usefulCount": {
            "type": "integer",
            "format": "int32"
          },
          "notUsefulCount": {
            "type": "integer",
            "format": "int32"
          },
          "viewCount": {
            "type": "integer",
            "format": "int32"
          },
          "myReviewType": {
            "$ref": "#/components/schemas/NullableOfDocumentReviewType"
          },
          "progress": {
            "$ref": "#/components/schemas/DocumentFileProgressDto"
          }
        }
      },
      "DocumentFileProgressDto": {
        "type": "object",
        "properties": {
          "documentFileId": {
            "type": "string",
            "format": "uuid"
          },
          "lastPage": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "lastAccessedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "nullable": true
      },
...
      "NullableOfDocumentReviewType": {
        "type": "integer",
        "nullable": true
      },
...
      "NullableOfVisibilityStatus": {
        "type": "integer",
        "nullable": true
      },
...
      "PagedResponseOfCommentDetailDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentDetailDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
...
      "PagedResponseOfDocumentDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentDto"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "hasPreviousPage": {
            "type": "boolean"
          },
          "hasNextPage": {
            "type": "boolean"
          }
        }
      },
...
      "SubjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "subjectName": {
            "type": "string"
          },
          "subjectCode": {
            "type": "string"
          }
        },
        "nullable": true
      },
...
      "TagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tagName": {
            "type": "string"
          }
        }
      },
...
      "TypeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "typeName": {
            "type": "string"
          }
        }
      },
...
      "UpdateCommentCommandRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
      "UpdateDocumentCommandRequest": {
        "type": "object",
        "properties": {
          "documentName": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "subjectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "typeId": {
            "type": "string",
            "format": "uuid"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "tagNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "authorIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "authors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuthorInput"
            },
            "nullable": true
          },
          "visibility": {
            "$ref": "#/components/schemas/NullableOfVisibilityStatus"
          },
          "fileIdsToRemove": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "nullable": true
          },
          "coverFileId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
...
      "VisibilityStatus": {
        "type": "integer"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT token. If not provided, token will be read from cookies.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}